requirements.txt
gunicorn>=19.4.5,<20.0
oauthlib>=2.0,<3.0
pyTelegramBotAPI>=3.6,<4.0
python-dotenv>=0.8,<1.0
SQLAlchemy>=1.1.12,<2.0
webapp2>=2.5.2,<3.0
WebOb>=1.6,<2.0
